feat: using function mappings in extension set
